pyproject.toml
[project]
name = "omnitower"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.13.0",
    "cassandra-driver>=3.29.3",
    "fastapi[standard]>=0.135.1",
]
